VNR-83-iOS-리뷰-등록-화면-UI-Autolayout-에러-수정

+++ Vinyler/Sources/Networking/Request/ReviewRequest.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 public struct ReviewRequest: Encodable {
-    let discogsId: Int64
+    let discogsId: Int
     let rating: Int
     let content: String?
 }
